requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==21.12b0
boto3==1.20.31
botocore==1.23.31
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
crispy-forms-materialize==0.2
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-allauth==0.47.0
django-cleanup==6.0.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-mathfilters==1.0.0
django-storages==1.12.3
gunicorn==20.1.0
heroku3==5.1.4
idna==3.3
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
sqlparse==0.4.2
stripe==2.65.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
